web/requirements.txt

gradio
requests
sseclient-py
pandas
